pyproject.toml
[project]
name = "predicting-customer-churn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.28",
    "click>=8.2.1",
    "fastapi>=0.115.12",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "mlflow>=2.22.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "xgboost>=3.0.1",
    "ydata-profiling>=4.16.1",
]
